requirements.txt
uvicorn==0.18.2
fastapi==0.78.0
python-multipart==0.0.5
torch==1.11.0
torchvision==0.12.0
Pillow==9.0.1